requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
pymongo==4.5.0
gunicorn==21.2.0
python-decouple==3.8
razorpay==1.4.2
